Outsource Device and VirtualDevice
